GraphRag/QueryWebap/requirements.txt
devtools 
pandas 
python-magic 
requests 
tqdm
streamlit
python-magic-bin